requirements.txt
asgiref==3.8.1
attrs==25.3.0
cffi==1.17.1
colorama==0.4.6
cryptography==44.0.1
Django==5.1.6
django-admin-sortable2==2.2.4
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==3.2.3
django-filter==25.1
django-formtools==2.5.1
django-grappelli==4.0.1
django-otp==1.5.4
django-phonenumber-field==8.0.0
django-two-factor-auth==1.17.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
et_xmlfile==2.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mysqlclient==2.2.7
openpyxl==3.1.5
phonenumbers==9.0.1
pycparser==2.22
PyMySQL==1.1.1
pypng==0.20220715.0
PyYAML==6.0.2
qrcode==7.4.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
